+++ Explorer++/Helper/Helper.h
@@ -30,8 +30,8 @@ BOOL			CreateSystemTimeString(const SYSTEMTIME *localSystemTime, TCHAR *szBuffer
 BOOL			CreateFriendlySystemTimeString(const SYSTEMTIME *localSystemTime, TCHAR *szBuffer, size_t cchMax);
 BOOL			GetFileSizeEx(const TCHAR *szFileName, PLARGE_INTEGER lpFileSize);
 BOOL			CompareFileTypes(const TCHAR *pszFile1,const TCHAR *pszFile2);
-HRESULT			BuildFileAttributeString(const TCHAR *lpszFileName, TCHAR *szOutput, DWORD cchMax);
-HRESULT			BuildFileAttributeString(DWORD dwFileAttributes, TCHAR *szOutput, DWORD cchMax);
+HRESULT			BuildFileAttributeString(const TCHAR *lpszFileName, TCHAR *szOutput, size_t cchMax);
+HRESULT			BuildFileAttributeString(DWORD dwFileAttributes, TCHAR *szOutput, size_t cchMax);
 BOOL			GetFileOwner(const TCHAR *szFile,TCHAR *szOwner,size_t cchMax);
 DWORD			GetNumFileHardLinks(const TCHAR *lpszFileName);
 BOOL			ReadImageProperty(const TCHAR *lpszImage, PROPID propId, TCHAR *szProperty, int cchMax);
